fix: cast state_attr to long before nn.Embedding lookup in EmbeddingBlock
MEGNet.predict_structure() builds state_attr with dtype float32 (matgl.float_th),
but nn.Embedding requires integer indices. Add .long() cast to match the existing
pattern used for node attributes when ntypes_node is set.

+++ src/matgl/layers/_embedding.py
@@ -103,7 +103,7 @@ class EmbeddingBlock(nn.Module):
             edge_feat = edge_attr
         if self.include_state is True:
             if self.ntypes_state and self.dim_state_embedding is not None:
-                state_feat = self.layer_state_embedding(state_attr)
+                state_feat = self.layer_state_embedding(state_attr.long())
             elif self.dim_state_feats is not None:
                 state_attr = torch.unsqueeze(state_attr, 0)
                 state_feat = self.layer_state_embedding(state_attr.to(matgl.float_th))
